requirements/base.txt
Django==1.9.5
djangorestframework==3.3.3
feedparser==5.2.1
python-dateutil==2.5.3
celery==3.1.23
xhtml2pdf

